% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_batch.R
\name{rpcr_batch}
\alias{rpcr_batch}
\title{A helper function to perform helper function in a batch mode}
\usage{
rpcr_batch(fun, df, id_col, ...)
}
\arguments{
\item{fun}{the helper function to be executed}

\item{df}{the data frame to be analysed}

\item{id_col}{the data column to be used}

\item{...}{addition arguments passed to fun}
}
\value{
returns the aggregated data calculated by fun
}
\description{
A helper function to perform helper function in a batch mode
}
