% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calAmp.R
\name{rpcr_calAmp}
\alias{rpcr_calAmp}
\title{detection of cycle range with significant amplification}
\usage{
rpcr_calAmp(
  df,
  start = "c",
  end = "f",
  zero = "c",
  cycle = TRUE,
  tres = 0.1,
  roff = c(0, 0)
)
}
\arguments{
\item{df}{data}

\item{start}{start cycle}

\item{end}{end cycle}

\item{zero}{zero value}

\item{cycle}{either TRUE or FALSE}

\item{tres}{threshold number form 0 to 1}

\item{roff}{cycle offset c(offset from start,offset from end)}
}
\value{
returns the range
}
\description{
detection of cycle range with significant amplification
}
\examples{
# currently no example
}
