% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calDer.R
\name{rpcr_calDer}
\alias{rpcr_calDer}
\title{derivative calculation}
\usage{
rpcr_calDer(df, offset = 1, ...)
}
\arguments{
\item{df}{data for derivative calculation}

\item{offset}{adding zero values}

\item{...}{further arguments, ignored}
}
\value{
derivative data
}
\description{
derivative calculation
}
