% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calEff.R
\name{rpcr_calEff}
\alias{rpcr_calEff}
\title{Efficiency calculation}
\usage{
rpcr_calEff(df, ...)
}
\arguments{
\item{df}{numeric vector of amplification data}

\item{...}{further arguments, ignored (accepted so that the function can
be called through rpcr_batch(), which passes id_col to every worker)}
}
\value{
efficiency data
}
\description{
Efficiency calculation
}
