% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calib.R
\name{rpcr_calib}
\alias{rpcr_calib}
\title{Calibrates rPCR data from calibration samples}
\usage{
rpcr_calib(df, prop = TRUE, extend = FALSE, novel = FALSE, silent = FALSE)
}
\arguments{
\item{df}{data from calibration samples}

\item{prop}{if TRUE data is in proportion, if FALSE in percentage}

\item{extend}{additional model}

\item{novel}{if novel = TRUE uses alternative model}

\item{silent}{if FALSE displays calibration error}
}
\value{
calibration parameters
}
\description{
Calibrates rPCR data from calibration samples
}
