% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_dpcr.R
\name{rpcr_dpeak}
\alias{rpcr_dpeak}
\title{Function to detect density peaks for the identification of cluster}
\usage{
rpcr_dpeak(
  df,
  bw = NULL,
  filter = NULL,
  plot = FALSE,
  scale = 100,
  output = ""
)
}
\arguments{
\item{df}{dPCR data}

\item{bw}{bandwith for density}

\item{filter}{filter}

\item{plot}{TRUE for plotting}

\item{scale}{rescaling data}

\item{output}{specifies output type}
}
\value{
peak information
}
\description{
Function to detect density peaks for the identification of cluster
}
