% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_filter.R
\name{rpcr_filter}
\alias{rpcr_filter}
\title{Savitzky-Golay smoothing of an amplification curve}
\usage{
rpcr_filter(df, ...)
}
\arguments{
\item{df}{input data}

\item{...}{further arguments, ignored (accepted so that the function can
be called through rpcr_batch(), which passes id_col to every worker)}
}
\value{
filtered data
}
\description{
Savitzky-Golay smoothing of an amplification curve
}
