% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_import.R
\name{rpcr_import}
\alias{rpcr_import}
\title{Imports a ThermoFisher PCR export file}
\usage{
rpcr_import(file)
}
\arguments{
\item{file}{file to be imported}
}
\value{
imported data
}
\description{
Imports a ThermoFisher PCR export file
}
