% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_inverse.R
\name{rpcr_inverse}
\alias{rpcr_inverse}
\title{Corrects proportion using the submitted parameters}
\usage{
rpcr_inverse(y, a, ri, re, g, f = 1, lower = -0.1, upper = 1.1, ...)
}
\arguments{
\item{y}{model parameter}

\item{a}{model parameter}

\item{ri}{model parameter}

\item{re}{model parameter}

\item{g}{model parameter}

\item{f}{model parameter}

\item{lower}{lower end point for uniroot}

\item{upper}{upper end point for uniroot}

\item{...}{passing arguments}
}
\value{
corrected ratio
}
\description{
Corrects proportion using the submitted parameters
}
