% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_norm.R
\name{rpcr_norm}
\alias{rpcr_norm}
\title{Title}
\usage{
rpcr_norm(df, type = c("T", "A"), mode = 2)
}
\arguments{
\item{df}{data to normalize}

\item{type}{???}

\item{mode}{???}
}
\value{
normalized data frame ?
}
\description{
Title
}
