% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_norm.R
\name{rpcr_normfit}
\alias{rpcr_normfit}
\title{Title}
\usage{
rpcr_normfit(df, detail = FALSE)
}
\arguments{
\item{df}{data to fit}

\item{detail}{TRUE for more detailed results}
}
\value{
coefficients or detailed fit
}
\description{
Title
}
