% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotCalib.R
\name{rpcr_plotCalib}
\alias{rpcr_plotCalib}
\title{Generating plot to validate calibration}
\usage{
rpcr_plotCalib(
  data = NULL,
  error = FALSE,
  ctype = "fix",
  confidence = FALSE,
  signal = "Rn",
  derivate = FALSE,
  percentage = FALSE,
  limit = c(-0.5, 0.5),
  calibrate = FALSE,
  para = NULL,
  smooth = FALSE,
  ratio = 2,
  color_target = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02",
    "#A6761D", "#666666"),
  color_sample = colorRampPalette(c("grey90", "grey30")),
  raw = TRUE,
  var = "aprop",
  ...
)
}
\arguments{
\item{data}{data}

\item{error}{displays errorbars if TRUE}

\item{ctype}{cycle}

\item{confidence}{confidence interval}

\item{signal}{type of signal}

\item{derivate}{derivate (default = FALSE)}

\item{percentage}{use derivative of signal when TRUE}

\item{limit}{axis limits of plots}

\item{calibrate}{perform calibration (default = FALSE)}

\item{para}{parameter for calibration}

\item{smooth}{smooth date}

\item{ratio}{Ratio between plots}

\item{color_target}{colors for target}

\item{color_sample}{colors for samples}

\item{raw}{raw ???}

\item{var}{selected variable}

\item{...}{arguments for passing}
}
\value{
plots
}
\description{
Generating plot to validate calibration
}
