% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotCalibCycle.R
\name{rpcr_plotCalibCycle}
\alias{rpcr_plotCalibCycle}
\title{Title}
\usage{
rpcr_plotCalibCycle(
  df,
  title = "",
  colplot = TRUE,
  box = TRUE,
  brew_y = brewer.pal(n = 8, name = "Dark2"),
  var = "cycle"
)
}
\arguments{
\item{df}{data  to be plotted}

\item{title}{alternative title text}

\item{colplot}{plotting in color (default = TRUE)}

\item{box}{plot a boxplot (defaulr = TRUE)}

\item{brew_y}{colors}

\item{var}{column to be plotted}
}
\value{
plots
}
\description{
Title
}
