% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotCalibQC.R
\name{rpcr_plotCalibQC}
\alias{rpcr_plotCalibQC}
\title{Title}
\usage{
rpcr_plotCalibQC(
  data,
  violin = TRUE,
  boxplot = TRUE,
  point = TRUE,
  bar = FALSE,
  abs = FALSE,
  color = brewer.pal(n = 8, name = "Dark2")
)
}
\arguments{
\item{data}{data to be plotted}

\item{violin}{plot violoin (default TRUE)}

\item{boxplot}{plot boxplot (default TRUE)}

\item{point}{plot data points (default TRUE)}

\item{bar}{plot error bars (default FALSE)}

\item{abs}{???}

\item{color}{plotting colors}
}
\value{
plots
}
\description{
Title
}
