% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotCalibSummary.R
\name{rpcr_plotCalibSummary}
\alias{rpcr_plotCalibSummary}
\title{pots a series of charts to summarize a calibration experiment using the functions rpcr_plotData and rpcr_plotBar}
\usage{
rpcr_plotCalibSummary(data, type = "all", signal = "all", ...)
}
\arguments{
\item{data}{data for plotting}

\item{type}{plot type}

\item{signal}{signal type}

\item{...}{passing arguments}
}
\value{
returns a list of plots
}
\description{
pots a series of charts to summarize a calibration experiment using the functions rpcr_plotData and rpcr_plotBar
}
