% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotPanel.R
\name{rpcr_plotPanel}
\alias{rpcr_plotPanel}
\title{Title}
\usage{
rpcr_plotPanel(df, wline = 0.5, slabel = 6, stitle = 8, ...)
}
\arguments{
\item{df}{data}

\item{wline}{line size}

\item{slabel}{label size}

\item{stitle}{title font size}

\item{...}{passing arguments}
}
\value{
plots
}
\description{
Title
}
