% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotPsi.R
\name{rpcr_plotPsi}
\alias{rpcr_plotPsi}
\title{rpcr_plotPsi}
\usage{
rpcr_plotPsi(
  para = c(a = 3, ri = 0.1, re = 0.2, g = 2),
  data = NULL,
  ltype = "solid",
  asize = 1.5,
  bsize = 1.5,
  csize = 0.5,
  ctype = "longdash",
  acol = "black",
  ccol = "grey30",
  balpha = 0.2,
  single = TRUE,
  sub = FALSE
)
}
\arguments{
\item{para}{parameter}

\item{data}{data to be ploted}

\item{ltype}{line type}

\item{asize}{line size}

\item{bsize}{line size}

\item{csize}{line size}

\item{ctype}{line type}

\item{acol}{color}

\item{ccol}{color}

\item{balpha}{alpha level}

\item{single}{???}

\item{sub}{===}
}
\value{
plots
}
\description{
rpcr_plotPsi
}
