% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_prop.R
\name{rpcr_prop}
\alias{rpcr_prop}
\title{Calibrate apparent proportions}
\usage{
rpcr_prop(
  aprop,
  calib = c(1, 0, 0, 1, 0),
  lower = -0.1,
  upper = 1.1,
  prop = TRUE,
  extend = FALSE,
  ...
)
}
\arguments{
\item{aprop}{apparent proportions to be calibrated}

\item{calib}{calibration parameters as a named vector, e.g.
c(a = 1, ri = 0, re = 0, g = 1)}

\item{lower}{lower limit passed to rpcr_inverse()}

\item{upper}{upper limit passed to rpcr_inverse()}

\item{prop}{proportion}

\item{extend}{extended model}

\item{...}{passing arguments}
}
\value{
calibrated proportions
}
\description{
Calibrate apparent proportions
}
\examples{
rpcr_prop(c(0.4, 0.2), calib = c(a = 3, ri = 0, re = 0.05, g = 1))
}
