% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_readTF.R
\name{rpcr_readTF}
\alias{rpcr_readTF}
\title{Reading ThermoFisher Export files}
\usage{
rpcr_readTF(filepath)
}
\arguments{
\item{filepath}{Optional character string specifying the path to the file.
If missing, a file selection dialog will open automatically.}
}
\value{
A data frame with extracted and standardized PCR amplification data.
}
\description{
Reads the amplification data sheet from ThermoFisher export files.
Auto-detects the header line and standardizes well names.
}
