% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rpcr.R
\name{scale_colour_rpcr}
\alias{scale_colour_rpcr}
\title{Discrete colour scale with the ratioPCR palette}
\usage{
scale_colour_rpcr(...)
}
\arguments{
\item{...}{further arguments passed to ggplot2::discrete_scale()}
}
\value{
colour scale
}
\description{
Discrete colour scale with the ratioPCR palette
}
