% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rpcr.R
\name{scale_fill_rpcr}
\alias{scale_fill_rpcr}
\title{Discrete fill scale with the ratioPCR palette}
\usage{
scale_fill_rpcr(...)
}
\arguments{
\item{...}{further arguments passed to ggplot2::discrete_scale()}
}
\value{
fill color scale
}
\description{
Discrete fill scale with the ratioPCR palette
}
