% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPCR-package.R
\docType{package}
\name{ratioPCR-package}
\alias{ratioPCR}
\alias{ratioPCR-package}
\title{ratioPCR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
