% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_batchFit.R
\name{rpcr_batchFit}
\alias{rpcr_batchFit}
\title{Function to calculate USI form multiple PCR experiments}
\usage{
rpcr_batchFit(data, calib = c(1, 0, 0, 1, 0), info = FALSE, signal = "Rn", ...)
}
\arguments{
\item{data}{data.frame with the pcr flourecence intensities}

\item{calib}{Calibration parameters to calibrate the calculted ratio}

\item{info}{if TRUE more detailed results}

\item{signal}{"Rn" for normalized Fluorescence values, anything else for baseline substracted values}

\item{...}{other parameters for rpcr_fit function}
}
\value{
returns fittings for calibration curves
}
\description{
Function to calculate USI form multiple PCR experiments
}
