% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calBase.R
\name{rpcr_calBase}
\alias{rpcr_calBase}
\title{Baseline detection}
\usage{
rpcr_calBase(df, offset = 1, tres = 0, ...)
}
\arguments{
\item{df}{amplification data}

\item{offset}{offset for derivative calculation}

\item{tres}{treshold}

\item{...}{other parameters}
}
\value{
the corrected values
}
\description{
Baseline detection
}
