% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calCip.R
\name{rpcr_calCip}
\alias{rpcr_calCip}
\title{Amplification curve detection}
\usage{
rpcr_calCip(
  x,
  qc_tres = 1.15,
  detail = FALSE,
  len = 10,
  mincycle = 10,
  stringent = TRUE,
  lead = 3,
  ...
)
}
\arguments{
\item{x}{a vector of PCR amplification singlas}

\item{qc_tres}{treshold}

\item{detail}{output details}

\item{len}{length parameter}

\item{mincycle}{minimum cycle}

\item{stringent}{stringency}

\item{lead}{lead cycles}

\item{...}{passing paramaters}
}
\value{
returns different amplification curve indicators
}
\description{
`rpcr_calCip` analyzes the amplification curve and returns starting, inflection and ending cycle of the amplification
}
\details{
Ther is no description for this function yet
}
