% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calEff.R
\name{rpcr_calEff}
\alias{rpcr_calEff}
\title{Efficiency calculation}
\usage{
rpcr_calEff(df, ...)
}
\arguments{
\item{df}{vector eith amplification data}
}
\value{
efficiency data
}
\description{
Efficiency calculation
}
