% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calibDetect.R
\name{rpcr_calibDetect}
\alias{rpcr_calibDetect}
\title{Detect calibrations samples}
\usage{
rpcr_calibDetect(data, qual = "Cal", percentage = FALSE, ...)
}
\arguments{
\item{data}{amplification data}

\item{qual}{qualifier for the calibration samples}

\item{percentage}{percentage (default = FALSE)}
}
\value{
amplification data with tPCI column for detected calibration samples
}
\description{
Detect calibrations samples
}
