% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calibQC.R
\name{rpcr_calibQC}
\alias{rpcr_calibQC}
\title{Testing quality of calibration experiment}
\usage{
rpcr_calibQC(df, tprop, ...)
}
\arguments{
\item{df}{PCR data}

\item{tprop}{theoretical proportions used to calibrate the data}

\item{...}{paramters for the rpcr_batchFit, rpcr_fit and rpcr_calib routines}
}
\value{
QC parameters
}
\description{
Testing quality of calibration experiment
}
