% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_calibrate.R
\name{rpcr_calibrate}
\alias{rpcr_calibrate}
\title{Title}
\usage{
rpcr_calibrate(
  data,
  output = "calib",
  median = FALSE,
  prop = TRUE,
  extend = FALSE,
  roff,
  tres,
  method,
  model,
  qual = "Cal",
  signal,
  rtype,
  ...
)
}
\arguments{
\item{data}{input data for calibration}

\item{output}{type of data to be returned}

\item{median}{median}

\item{prop}{proportion setting}

\item{extend}{extended model}

\item{roff}{relative cycle offset}

\item{tres}{treshold}

\item{method}{method}

\item{model}{model}

\item{qual}{calibration samples qualifier}

\item{signal}{signal type}

\item{rtype}{range type}

\item{...}{parameters passed to other functions}
}
\value{
calibration parameters
}
\description{
Title
}
