% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_fit.R
\name{rpcr_fit}
\alias{rpcr_fit}
\title{Calculate USI from splice PCR data}
\usage{
rpcr_fit(
  df,
  tres = c(0, 0),
  rtype = "ip",
  roff = c(0, 0),
  model = "linear",
  method = c("FA", "FT"),
  calib = c(1, 0, 0, 1),
  plot = FALSE,
  average = TRUE,
  offset = 1,
  derivative = FALSE,
  baseline = FALSE,
  ...
)
}
\arguments{
\item{df}{Data frame containing the flourescence data of the inclusion and exclusion probe}

\item{tres}{Threshold value to define a range of cycles used for USI calculation when the range type is dynamic.}

\item{rtype}{A vector of two values defining the method, how to determine the starting and ending cycle used for USI calculation
Possible methods c("dynamic","full","ip",c(1:dim((df))))}

\item{roff}{offset from range defined by rtype}

\item{model}{Single String defining the model used to calculate the USI c("linear","single")}

\item{method}{Method type how to calculate the ratio}

\item{calib}{calibration parameter for ratio correction}

\item{plot}{print plot (default = FALSE)}

\item{average}{default = TRUE}

\item{offset}{offset for derivate calculation}

\item{derivative}{data transformed to 1. derivative (default = FALSE)}

\item{baseline}{baseline correction (default = FALSE)}

\item{...}{parameters passed to other functions}
}
\value{
calculated ratio
}
\description{
Calculate the USI from splice PCR data. Inclusion and exclusion probe data.
Different methods can by used for the calculation of USI
}
