% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_fitPlot.R
\name{rpcr_fitPlot}
\alias{rpcr_fitPlot}
\title{Title}
\usage{
rpcr_fitPlot(
  x,
  offset = 1,
  design = theme_rpcr(),
  iso_color = c(grDevices::rgb(199, 235, 211, maxColorValue = 255), grDevices::rgb(255,
    190, 190, maxColorValue = 255)),
  aslist = FALSE,
  line = FALSE,
  ...
)
}
\arguments{
\item{x}{data}

\item{offset}{offset}

\item{design}{design}

\item{iso_color}{colors}

\item{aslist}{default = FALSE}

\item{line}{draw line (default = FALSE)}

\item{...}{passing arguments}
}
\value{
generated plots
}
\description{
Title
}
