% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotBar.R
\name{rpcr_plotBar}
\alias{rpcr_plotBar}
\title{Plotting Bar charts of calibration Data}
\usage{
rpcr_plotBar(
  data,
  cycle = NULL,
  ctype = "fix",
  signal = "Rn",
  derivate = "",
  title = "normalized flourescence\\n ",
  color_target = brewer.pal(n = 8, name = "Dark2"),
  color_sample = colorRampPalette(c("grey90", "grey30")),
  colplot = FALSE,
  ...
)
}
\arguments{
\item{data}{data}

\item{cycle}{???}

\item{ctype}{???}

\item{signal}{type of signal}

\item{derivate}{derivate}

\item{title}{alternative plot title}

\item{color_target}{color vector for different targets}

\item{color_sample}{color vector for different samples}

\item{colplot}{TRUE id plot in color}

\item{...}{passing arguments}
}
\value{
plots
}
\description{
Plotting Bar charts of calibration Data
}
