% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotData.R
\name{rpcr_plotData}
\alias{rpcr_plotData}
\title{Title}
\usage{
rpcr_plotData(
  df0,
  signal = "Rn",
  derivate = "",
  title = "normalized flourescence\\n ",
  color_target = RColorBrewer::brewer.pal(n = 8, name = "Dark2"),
  color_sample = grDevices::colorRampPalette(c("grey90", "grey30")),
  color_probe = c(rgb(199, 235, 211, maxColorValue = 255), rgb(255, 255, 190,
    maxColorValue = 255), rgb(255, 190, 190, maxColorValue = 255)),
  colplot = FALSE,
  label = TRUE,
  wline = 0.5,
  slabel = 6,
  stitle = 8,
  text = TRUE,
  alpha = NULL,
  targetA = "target A",
  targetB = "target B",
  ...
)
}
\arguments{
\item{df0}{data}

\item{signal}{type of signal}

\item{title}{alternative title text}

\item{color_target}{color of targets}

\item{color_sample}{color of samples}

\item{colplot}{TRUE for color plotting}

\item{...}{passing arguments}
}
\value{
plots
}
\description{
Title
}
