% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_plotSim.R
\name{rpcr_plotSim}
\alias{rpcr_plotSim}
\title{Title}
\usage{
rpcr_plotSim(
  para = NULL,
  ltype = "solid",
  asize = 1.5,
  bsize = 1.5,
  csize = 0.5,
  ctype = "longdash",
  acol = "black",
  ccol = "grey30",
  balpha = 0.2,
  append = NULL,
  ratio = 2,
  single = FALSE,
  lincols = brewer.pal(n = 8, name = "Dark2"),
  ...
)
}
\arguments{
\item{para}{parameter for simulation}

\item{ltype}{line type}

\item{asize}{line size}

\item{bsize}{line size}

\item{csize}{line size}

\item{ctype}{line type}

\item{acol}{color}

\item{ccol}{color}

\item{balpha}{alpha}
}
\value{
plot
}
\description{
Title
}
