% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_prop.R
\name{rpcr_prop}
\alias{rpcr_prop}
\title{Title}
\usage{
rpcr_prop(
  aprop,
  calib = c(1, 0, 0, 1, 0),
  lower = -0.1,
  upper = 1.1,
  prop = TRUE,
  extend = FALSE,
  ...
)
}
\arguments{
\item{aprop}{apperent ratios to be calibrated}

\item{calib}{Calibration parameters as named list like c(a = 1, ri = 0, re = 0)}

\item{lower}{lower limits}

\item{upper}{upper limits}

\item{prop}{proportion}

\item{extend}{extended model}

\item{...}{passing arguments}

\item{df}{data}
}
\value{
calibrated ratio
}
\description{
Title
}
\examples{
rpcr_prop(df = data.frame(0.4, 0.2), calib = c(a = 3, ri = 0, re = 0.05))
}
