% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_readTF.R
\name{rpcr_readTF}
\alias{rpcr_readTF}
\title{Reading ThermoFisher Export files}
\usage{
rpcr_readTF(filepath)
}
\arguments{
\item{filepath}{optional specifiy export file}
}
\value{
data frame with pcr data and calculations
}
\description{
Reading ThermoFisher Export files
}
