% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_simprop.R
\name{rpcr_simprop}
\alias{rpcr_simprop}
\title{Function to simulate ratioPCR}
\usage{
rpcr_simprop(
  tprop = c(0, 0.05, 0.15, 0.25, 0.35, 0.5, 0.65, 0.75, 0.85, 0.95, 1),
  a = 1,
  ri = 0,
  re = 0,
  g = 1,
  f = 0.5,
  prop = TRUE,
  extend = FALSE,
  novel = FALSE,
  ...
)
}
\arguments{
\item{tprop}{a theoretical proportion value or a list of proportion values}

\item{a}{The alpha (a) value describes the difference in
probe intensities (incl/excl)}

\item{ri}{The ratio ri describes the cross-reactivity of the intensity probe

The ratio can be between 0 and 1.
ri = 0 -> The inclusion probe is specific there is no cross reactivity with the exclusion isoform
ri = 1 -> The inclsuion probe is unspecific. The probe recognizes both isoforms equally.
0 > ri > 1 -> There is a certain specificity towards the inclusion isoform, but also recognizes the exclusion isoform.}

\item{re}{The ratio re describes the cross-reactivity of the exclusion probe

The ratio can be between 0 and 1.
re = 0 -> The exclusion probe is specific there is no cross reactivity with the inclusion isoform
re = 1 -> The exclusion probe is unspecific. The probe recognizes both isoforms equally.
0 > ri > 1 -> There is a certain specificity towards the exclusion isoform, but also recognizes the inclusion isoform.}

\item{g}{parameter}
}
\value{
The simulated biased PSI
}
\description{
Function to simulate ratioPCR
}
