% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_wrap.R
\name{rpcr_wrap}
\alias{rpcr_wrap}
\title{Reformatting of imported data and additional calculations}
\usage{
rpcr_wrap(df, full = FALSE, convert = NULL)
}
\arguments{
\item{df}{data frame with the PCR data}

\item{full}{TRUE For additional calculations}

\item{convert}{???}
}
\value{
data frame with results ?
}
\description{
Reformatting of imported data and additional calculations
}
