% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcr_settings.R, R/theme_rpcr.R
\name{theme_rpcr}
\alias{theme_rpcr}
\title{Title}
\usage{
theme_rpcr(base_size = 10, base_family = "sans")

theme_rpcr(base_size = 10, base_family = "sans")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{
theme for plotting

theme for ggplot
}
\description{
Title

theme for ggplot figures
}
